requirements.txt
tensorflow>=2.12
numpy
pandas
matplotlib
seaborn
scikit-learn
opencv-python
Pillow
tqdm
Flask>=2.3.0
Werkzeug>=2.3.0
retry>=0.9.2
